+++ Assets/VRM10/Editor/Components/SpringBone/SelectedGUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3afd3f6a158933546aa5d979acdf7030
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
